app/requirements.txt
fastapi
uvicorn
databases
# psycopg2
psycopg2-binary
fastapi_mail